pyproject.toml
[project]
name = "poster-generator"
version = "1.1.0"
description = ""
authors = [
    {name = "corgi-in-tights",email = "59304120+corgi-in-tights@users.noreply.github.com"}
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "pillow (>=12.0.0,<13.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[project.urls]
homepage = "https://github.com/corgi-in-tights/poster-generator"
repository = "https://github.com/corgi-in-tights/poster-generator"

[tool.setuptools.packages.find]
include = ["poster_generator", "poster_generator.*"]

[tool.poetry.group.dev.dependencies]
twine = "^6.2.0"
build = "^1.3.0"
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# ==== ruff ====
[tool.ruff]
line-length = 119
target-version = "py312"
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F", "E", "W", "C90", "I", "N", "UP", "YTT",
  "ASYNC", "S", "BLE", "FBT", "B", "A", "COM", "C4",
  "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN",
  "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET",
  "SLF", "SLOT", "SIM", "TID", "TC", "INT", "PTH", "ERA",
  "PD", "PGH", "PL", "TRY", "FLY", "PERF", "RUF"
]
ignore = [
  "S101",     # Use of assert detected
  "RUF012",   # Mutable class attributes without typing.ClassVar
  "SIM102",   # Allow nested ifs sometimes
  "UP038"     # isinstance tuple perf issue
]
extend-unsafe-fixes = ["UP038"]

[tool.ruff.lint.isort]
force-single-line = true